feat(frontend): implement dynamic grid and refactor games to use
[frontend] Add absolute positioned board

This board is build as a base for better game-generic features, while
also allowing for game specific overlays/function to be added.

[frontend] Add comments to useBoardLayout

[frontend] Remove unused Dockerfile.postgres

[frontend] Update Sudoku to use new board

[frontend] Fix typo, remove absolute from css

add position as class instead so it's clear without digging

[frontend] Update tents instructions

I said tent, then showed a tree. Fixed that.

Thanks for catching it @chentoast

[frontend] Add dynamic game grid

Check README.md for description on functionality and current issues.
I will refer to this as "GameGrid"

[frontend] Fix minesweeper correctness logic

[frontend] Switch all current games to use new dynamic game grid

Fixes #12

[frontend] Add board dev tools, organize main.ts+icons.ts+app.vue

Dev tools for making board easier to make changes to

[frontend] Add board.interaction.ts

Instead of the typical event handler (which could have performance
issues if used at the scale this could go to with each border + cell
haveing approx 9 event listeners on them), we use this interaction
layer, which will calculate what is below the mouse, and propogate the
event with that cell/border information

[frontend] npm update